Add files via upload
Applied algorithmic trading on Apple's shares data set. with the help of Numpy, pandas, and matplotlib libraries of Python Used Moving Average Convergence Divergence (MACD) to analyze the data and set the buying and selling point in the graph. with the help of exponential moving average (EMA) plotted the point so that the user can easily buy and sell the shares.